+++ blocks/markdown/src/index.ts
@@ -3,14 +3,10 @@ import marked from 'marked';
 
 import styles from './style.css';
 
-interface BlockParameters {
-  content: string;
-}
-
 /**
  * @param {Object} block The block as it was specified by the app creator.
  */
-attach<BlockParameters>(({ block }) => {
+attach(({ block }) => {
   const { content } = block.parameters;
 
   const markdown = document.createElement('div');